--- src/components/Body/AccessMyWallet/AccessMyWalletOptions/AccessMyWalletOptions.scss
@@ -1,350 +1,0 @@
-@import '~@/scss/GlobalVariables';
-
-// Bootstrap modal --------------------------------------------
-.modal-container {
-  .bootstrap-modal {
-    .modal-icon {
-      display: flex;
-      margin-bottom: 35px;
-      margin-top: 10px;
-
-      img {
-        margin: 0 auto;
-      }
-    }
-
-    .content-container {
-      margin-bottom: 20px;
-    }
-
-    .accept-terms {
-      margin-bottom: 20px;
-      text-align: center;
-
-      .container {
-        display: inline;
-      }
-    }
-
-    .button-container {
-      margin-bottom: 25px;
-
-      .connection-button {
-        cursor: initial;
-        text-align: center;
-        width: 100%;
-      }
-
-      .waiting-for-connection {
-        background-color: $button--background-color-3;
-
-      }
-    }
-  } // .bootstrap-modal
-
-  .appstore-button-container {
-    background-color: $background-color-6;
-    border-radius: 4px;
-    margin-bottom: 30px;
-    padding: 30px;
-    text-align: center;
-
-    p {
-      margin: 0;
-    }
-
-    img {
-      margin-bottom: 13px;
-      width: 130px;
-    }
-  }
-
-  .support {
-    display: flex;
-
-    a {
-      margin: 0 auto;
-    }
-
-    .support-content {
-      align-items: center;
-      display: flex;
-      margin: 0 auto;
-
-      .support-icon img {
-        height: 25px;
-        margin-right: 10px;
-      }
-
-      .support-label h5 {
-        color: $link-text-color-2;
-        font-size: 14px;
-      }
-    } // .support-content
-  } // .support
-
-  .modal-network-and-address {
-    .content-container-1 {
-      margin: -20px -40px 0;
-      padding: 0 25px;
-
-      .hd-derivation {
-        align-items: center;
-        display: flex;
-
-        .dropdown-button-container {
-          margin-left: auto;
-        }
-      }
-
-      .derivation-brands {
-        font-size: 12px;
-        padding: 15px 0;
-        text-align: left;
-      }
-    } // .content-container-1
-
-    .content-container-2 {
-      background-color: $background-color-1;
-      margin: 0 -40px;
-      margin-bottom: 30px;
-      padding: 0;
-
-      .address-nav {
-        padding: 10px 0;
-        text-align: center;
-
-        span {
-          cursor: pointer;
-          padding: 2px 10px;
-        }
-      }
-
-      .address-block-container {
-        .block-title {
-          padding: 15px 25px;
-        }
-
-        .address-data:nth-child(even) {
-          background-color: $background-color-white;
-        }
-
-        .address-block {
-          color: $text-color-3;
-          display: grid;
-          grid-column-gap: 15px;
-          grid-template-columns: 35px 3fr 2fr 35px;
-          padding: 8px 25px;
-
-          &.table-header {
-            background-color: $background-color-4;
-            color: $text-base-color;
-            font-weight: 600;
-          }
-
-          &.selected {
-            color: $text-color-6;
-            font-weight: 500;
-          }
-
-          li {
-            overflow: hidden;
-            text-overflow: ellipsis;
-            white-space: nowrap;
-          }
-        }
-      } // .address-block-container
-    } // .content-container-2
-
-    .close-button,
-    .close-button * {
-      width: 100%;
-    }
-  } // .modal-mew-connect2
-
-  .modal-metamask {
-    .modal-multi-icons {
-      align-items: center;
-      display: flex;
-      justify-content: center;
-      margin-bottom: 25px;
-      margin-top: 20px;
-
-      img {
-        padding: 0 13px;
-      }
-
-      .logo-small {
-        max-height: 90px;
-      }
-    }
-
-    .content-container {
-      margin-bottom: 20px;
-
-      h4 {
-        font-weight: 400;
-        line-height: 22px;
-        margin: 0 auto 40px;
-        max-width: 340px;
-      }
-    }
-  }
-
-  .modal-hardware,
-  .modal-software {
-
-    .button-options {
-      display: grid;
-      grid-gap: 10px;
-      grid-template-columns: 1fr 1fr;
-      margin-bottom: 40px;
-
-      &.hardware-button-options {
-        margin-bottom: 30px;
-      }
-
-      li {
-        align-items: center;
-        color: $text--link-color;
-        cursor: pointer;
-        display: flex;
-        padding: 30px 0 30px 30px;
-        text-align: left;
-        user-select: none;
-
-        * {
-          pointer-events: none;
-        }
-
-        span {
-          line-height: 16px;
-        }
-
-        &.active {
-          box-shadow: 0 3px 15px $box-shadow-color-1;
-          color: $text-black;
-          font-weight: 500;
-
-          .icon {
-            display: none;
-          }
-
-          .icon-hover {
-            display: block;
-          }
-        }
-
-        &:hover {
-          color: $text-black;
-          font-weight: 500;
-
-          .icon {
-            display: none;
-          }
-
-          .icon-hover {
-            display: block;
-          }
-        }
-
-        .icon,
-        .icon-hover {
-          margin-right: 8px;
-          width: 50px;
-        }
-
-        .icon-hover {
-          display: none;
-        }
-
-      }
-    } // .button-options
-  } // .modal-software .modal-hardware
-} // .modal-container
-
-
-
-
-.wrap {
-  background-color: $background-color-1;
-  padding: 80px 0;
-}
-
-.title {
-  margin-bottom: 50px;
-  text-align: center;
-}
-
-.buttons-container {
-  display: grid;
-  grid-column-gap: 20px;
-  grid-template-columns: 1fr 1fr 1fr 1fr;
-  margin: 0 auto;
-  max-width: 900px;
-  position: relative;
-  text-align: center;
-
-  .button-block {
-    background-color: $background-color-white;
-    border-radius: 5px;
-    cursor: pointer;
-    padding: 10px 15px 25px;
-    position: relative;
-    top: 0;
-    transition: all .3s ease;
-
-    .button-image {
-      font-size: 90px;
-      margin-bottom: 5px;
-    }
-
-    p {
-      line-height: 20px;
-    }
-
-    .small-note {
-      color: $link-text-color-2;
-      font-size: 11px;
-      margin-top: 5px;
-    }
-
-    h3 {
-      margin-bottom: 10px;
-    }
-
-    .tooltip-box {
-      display: none;
-      position: absolute;
-      right: 15px;
-      top: 10px;
-
-      .btn {
-        background-color: transparent;
-        border: 0;
-        color: $text-white;
-        font-size: 18px;
-        line-height: 100%;
-        padding: 0;
-      }
-    }
-
-    &:hover {
-      background-color: $background-color-2;
-      box-shadow: 0 5px 20px $box-shadow-color-1;
-      color: $text-white;
-      top: -8px;
-
-      h3,
-      p {
-        color: $text-white;
-      }
-
-      .tooltip-box {
-        color: $link-text-color-2;
-        display: block;
-      }
-    }
-
-  }
-}
