--- src/components/Body/Transactions/SendOffline/SendOffline.scss
@@ -1,118 +1,0 @@
-@import '~@/scss/GlobalVariables';
-
-.wrap {
-  display: grid;
-  grid-column-gap: 0;
-  grid-template-columns: 270px 1fr;
-
-  .side-nav {
-    // background-color: yellow;
-  }
-
-  .contents {
-    background-color: $background-color-1;
-    padding: 40px 60px;
-
-    .tx-contents {
-      display: grid;
-      grid-gap: 15px;
-      grid-template-areas: 'info-1 info-2 info-3' 'main main tokens';
-      grid-template-columns: 1fr 1fr 1fr;
-      grid-template-rows: auto;
-
-      .main-content {
-        border-radius: 5px;
-        grid-area: main;
-        overflow: hidden;
-
-        >div {
-          background-color: $background-color-white;
-        }
-
-        $border-bottom-color: rgba(0, 0, 0, .06);
-
-        .content-title {
-          // box-shadow: 0 1px 1px $box-shadow-color-2;
-          border-bottom: 1px solid $border-bottom-color;
-          padding: 17px 0;
-
-          h2 {
-            border-left: 5px solid $background-color-2;
-            font-size: 24px;
-            font-weight: 500;
-            line-height: 37px;
-            margin: 0;
-            padding: 0;
-            padding-left: 27px;
-          }
-        }
-
-        .progress-status {
-          $border-bottom-black: #000;
-          background-color: $background-color-5;
-          border-bottom: 1px solid $border-bottom-color;
-          box-shadow: 0 2px 0 $border-bottom-black;
-          display: grid;
-          grid-template-columns: 27% 27% 27%;
-          text-align: center;
-          grid-column-gap: 20px;
-          justify-content: space-evenly;
-
-          >div {
-            padding-top: 24px;
-            padding-bottom: 18px;
-            cursor: pointer;
-
-            p {
-              display: block;
-              margin: 0 auto;
-              color: $text-color-9;
-            }
-
-            .title {
-              font-weight: 600;
-            }
-
-          }
-
-          .active {
-            border-bottom: 4px solid black;
-
-            p {
-              color: $text-black;
-            }
-          } 
-
-        }
-
-        .advanced {
-          padding-top: 10px;
-
-          .advanced-content {
-            border-bottom: 2px solid $background-color-1;
-            border-top: 2px solid $background-color-1;
-            display: flex;
-            padding: 0 8px;
-            padding-bottom: 20px;
-            padding-top: 20px;
-
-            .toggle-button {
-              align-items: center;
-              display: flex;
-              margin-left: auto;
-
-              >span {
-                margin-right: 15px;
-              }
-            }
-          }
-        } // .advanced
-
-      }
-
-      .tokens {
-        border-radius: 5px;
-      }
-    }
-  }
-}
