+++ src/components/GenerateInfo/GenerateInfo.vue
