+++ src/components/GenerateInfo/index.js
@@ -1,0 +1,1 @@
+export { default } from './GenerateInfo'
