+++ src/components/GenerateTx/GenerateTx.scss
@@ -16,10 +16,10 @@
           display: flex;
           margin-bottom: 2px;
 
-          h4 {                  
+          h4 {
             margin-right: 10px;
           }
-        }              
+        }
       }
 
       .buttons {
@@ -168,7 +168,7 @@
         display: flex;
         margin-left: auto;
 
-        >span {
+        > span {
           margin-right: 15px;
         }
       }
