+++ src/components/HardwareModal/HardwareModal.scss
@@ -7,6 +7,10 @@
     grid-template-columns: 1fr 1fr;
     margin-bottom: 40px;
 
+    &.hardware-button-options {
+      margin-bottom: 30px;
+    }
+
     li {
       align-items: center;
       color: $text--link-color;
@@ -14,13 +18,31 @@
       display: flex;
       padding: 30px 0 30px 30px;
       text-align: left;
+      user-select: none;
+
+      * {
+        pointer-events: none;
+      }
 
       span {
         line-height: 16px;
       }
 
-      &:hover {
+      &.active {
         box-shadow: 0 3px 15px $box-shadow-color-1;
+        color: $text-black;
+        font-weight: 500;
+
+        .icon {
+          display: none;
+        }
+
+        .icon-hover {
+          display: block;
+        }
+      }
+
+      &:hover {
         color: $text-black;
         font-weight: 500;
 
@@ -42,7 +64,6 @@
       .icon-hover {
         display: none;
       }
-
     }
   }
 }
